+++ client/includes/Plugin.php
@@ -55,6 +55,7 @@ final class Plugin {
 		add_action( 'admin_menu', array( Admin\TimelineScreen::class, 'register' ) );
 		add_action( 'admin_menu', array( Admin\CalendarScreen::class, 'register' ) );
 		add_action( 'admin_menu', array( Admin\AskScreen::class, 'register' ) );
+		add_action( 'admin_menu', array( Admin\AskedScreen::class, 'register' ) );
 		add_action( 'admin_menu', array( Admin\ConnectionScreen::class, 'register' ) );
 
 		Admin\AskActions::boot();